save renderer instructions to JSON in container
This commit adds a decorator that applies to instance methods on
Renderer objects and saved the instructions to a running list. At the
moment, registered templates are not added to the dictionary, but that
must be done before this is merged into the refactor branch. This commit
also adds a function that saves this JSON info to a file inside the
container to `/.reproenv.json`.