add json start+stop and maybe switch user for json
To ease testing (and perhaps readability of rendered artifacts), we
bookend the JSON saving with two comments. During testing, the content
between these two lines can be removed.

We also keep track of the current user in the _Renderer base class now.
This is necessary because when saving the JSON file, we need root
privileges. If we are not root, we switch to root temporarily to write
the JSON file.