do not rm docker images by default
Do not remove docker images by default in each test. Instead, use a
pytest hook to remove all docker images prepended with a certain string.
This hook is run at the end of the testing session.

Also add a function `get_build_and_run_fns` to help with tests
parameterized by builder. One can pass the builder used (docker or
singularity) and get the appropriate functions to build the image and to
run it.