fix gsl symlinking
The original code searched for a specific version of gsl. This does not
work because different distro versions can install different versions of
gsl. So instead, search for the present gsl version. Assumes that it
consists of two values.

This also fixes a whitespace issue, in which the symlinking code started
on the same line as an apt pruning command.