FIX afni installation (#401)
* use echo instead of printf for saving json

* fix types + error on None param name

* use printf for json + escape () and %

* fix gsl symlinking

The original code searched for a specific version of gsl. This does not
work because different distro versions can install different versions of
gsl. So instead, search for the present gsl version. Assumes that it
consists of two values.

This also fixes a whitespace issue, in which the symlinking code started
on the same line as an apt pruning command.

* require click < 8.0

Click version 8.0 seems to break our CLI... for example, the multi
option --install returns a string with click 8. Before, it would return
a tuple. I do not know why it would be returning a string...