remove /docs from .dockerignore
By excluding `/docs` from the Docker images, we were modifying the state
of the git repository, and `git status` was reporting uncommitted
changes. This caused the command `neurodocker --version` to show that
there were uncommitted changes (by including the `dirty` suffix in the
version). We do not want users to be under the impression that the
version includes uncommitted changes. Indeed, kaczmarj excluded `/docs`
from the docker builds as a premature optimization... it turns out that
they only consume 136 kb, which is negligible.