fix the types in --copy and --entrypoint
Both of these options use the class OptionEatAll. By default, this class
will output a string type, but that would result in a string that looked
like a tuple. For example, it is possible to use `ast.literal_eval(value)`
on the value to convert it back to a tuple. To fix this, this commit
sets the type for --copy and --entrypoint to `tuple` and the subsequent
code expects the values for --copy and --entrypoint to be tuples.

This commit also adds a type hint to an OptionsEatAll internal to help
with debugging.