raise an error if OptionEatAll is used as a str type
The fundamental issue leading to #498 was that the values in
OptionEatAll were being coerced to a string type if no type was
specified in OptionEatAll. This made a tuple `('foo',)` into a string
`'("foo",)'`. This commit raises an error if OptionEatAll is used as a
string. Indeed, the 'eat-all' nature of this option suggests that there
are multiple options, which suggests the type should be a list or tuple.

This commit also adds a test for this.