do not install sphinx 6.x
The pydata theme we are using seems to be incompatible with sphinx 6.x
which was released recently. We prevent this version from being
installed for the time being. The error produced with sphinx 6.x is
below and is related to
https://github.com/pydata/pydata-sphinx-theme/issues/1094.

This has been fixed in pydata-sphinx-theme but has not been released
yet. The fix should be part of the next release 0.13.

```
Theme error:
An error happened in rendering the page api.
Reason: UndefinedError("'logo' is undefined")
make: *** [Makefile:20: html] Error 2
```